Spring Security Working for Both User and Journal.

+++ src/main/java/net/engineeringdigest/journalApp/service/JournalEntryService.java
@@ -21,11 +21,17 @@ public class JournalEntryService {
 
     @Transactional
     public void saveEntry(JournalEntry journalEntry, String username) {
+        try {
+            User user = userService.findByUsername(username);
+            JournalEntry saved = journalEntryRepository.save(journalEntry);
+            user.getJournalEntries().add(saved);
+            userService.saveUser(user);
+        }catch(Exception e)
+        {
+            System.out.println(e);
+            throw new RuntimeException("An error occured while saving the entry" ,e);
+        }
 
-        User user = userService.findByUsername(username);
-        JournalEntry saved = journalEntryRepository.save(journalEntry);
-        user.getJournalEntries().add(saved) ;
-        userService.saveEntry(user);
     }
 
     public void saveEntry(JournalEntry journalEntry) {
@@ -41,12 +47,30 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String username){
-        User user = userService.findByUsername(username);
-        user.getJournalEntries().removeIf(x-> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String username)
+    {
+        boolean removed = false;
+        try {
+            User user = userService.findByUsername(username);
+             removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+            return removed;
+        }
+        catch(Exception e)
+        {
+            System.out.println(e);
+            throw new RuntimeException("An error occured while deleting the entry", e);
+        }
+
+
+
     }
+    //public List<JournalEntry> findByUsername(String Username)
+
 
 }
